Move ws jsonrpc headers to different file
